Card for minigames GoFP
