+++ pages/index.tsx
@@ -9,6 +9,7 @@ import { AWS_ASSETS_PATH } from '../src/constants'
 const assets = {
   airdrop: `${AWS_ASSETS_PATH}/airdrop.png`,
   terminus: `${AWS_ASSETS_PATH}/Terminus.png`,
+  minigames: `${AWS_ASSETS_PATH}/minigames-card.png`,
 }
 
 const features = [
@@ -23,6 +24,12 @@ const features = [
     description: 'Claim your token',
     image: assets.airdrop,
     href: '/airdrop',
+  },
+  {
+    name: 'Minigames',
+    description: 'Start your adventure at Garden of Forking Paths',
+    image: assets.minigames,
+    href: '/minigames',
   },
 ]
